connectors/common/pyproject.toml
[project]
name = "keelson-connectors-common"
version = "0.0.0"
description = "Shared utilities for Keelson connectors"
requires-python = ">=3.11"
dependencies = [
    "keelson",
    "eclipse-zenoh>=1.2.1",
]

[tool.uv.sources]
keelson = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/keelson_connectors_common"]
